refactor: create reusable task components
